Let breadcrumbs be scrollable
The trick is adding overflow in either direction creates a new box context, which means dopdowns can't escape. The solution is making a glass layer, where the nav is a child of an element that fills the whole screen and has overflow-x: auto. The part of the glass layer that overlaps the other content has pointer-events: none so you can still click through.