Fix scrollbar for wide breadcrumbs
If the breadcrumbs are wider than the current screen, the scrollbar appears at the bottom of the page. This is possible distracting on a puzzle page, but really bad on the blackboard, where you can scroll past the scrollbar but not interact with it because of pointer-events: none. This change hides that scrollbar. It will require an option setting in Firefox 63, but that option will become the default in Firefox 64, which releases in a few days.
Fixes Torgen#72