Add favorites field
Each puzzle now has a list of users who have favorited it. We use a $elemMatch field projection to filter the entries down to just yourself, both for relative privacy, not to push spurious updates over the wire, and to make filtering based on whether you've favorited a puzzle faster. Unfortunately since $elemMatch counts as an inclusion, not an exclusion, we have to explicitly enumerate every field in the Puzzles collection. Hopefully I didn't miss any. Alternatively I could exclude favorites from the regular publish and add a manual publish where I add it; this is what I would have to do if I wanted to push a favorite count without pushing the whole list.