Increase parallelism in drive client
When creating a new puzzle, create the spreadsheet and doc in parallel. Also, start creating the spreadsheet and doc while waiting to set the permissions on the folder.
When setting permissions on an object, set them all in parallel.
When renaming a puzzle, rename the folder, the spreadsheet, and the doc in parallel.
Also removes apiThrottle, since the modern Google APIs client does this for us.