Eliminate setPresence method
Instead you register your presence with a subscription. When you change rooms, you subscribe for the new room, then unsubscribe from the old one; when you close the tab or go to a non-chat view, you unsubscribe as well. In both cases we update the database to show you as not being in the room on that connection. This lets us generate the part message immediately when you leave from all clients, without spuriously generating it if you leave from one client but are still present in another.
For the time being this means we're not generating the foreground signal, so you always show up as being in the background.