Put gravatars in mentions and tags
Instead of converting messages to a block of HTML, convert them to a list of parts and render the parts with templates.
Also fixes the issue where the url of an image in a tag would embed the image in the blackboard but not the puzzle tags. Now it doesn't embed it anywhere.
Unfortunately required reimplementing Template.dynamic without extraneous whitespace.