Fix race condition on deleteRound
It was possible to create a new puzzle in a round at the same time as someone was deleting the round, which would leave the puzzle in the database with no way of seeing it. This prevents that by having the check for the round being empty happen as part of the remove call.