Drag and drop link to make puzzle/round
Dropping a link onto the New Round button makes a round. Dropping a link onto a round name under the New Meta button makes a new meta in that round. Dropping a link onto a round name under the New Puzzle button make a standalone puzzle in that round. Dropping a link on an existing meta adds a new feeder to that meta. In all cases, the link to the new puzzle is the href of the link being dragged, and the name is the text of the link.