Fix race condition on presence index (#1126)
If you leave and rejoin a room at just the right time, the added and
removed messages can get delivered out of order, which meant you were
added to a room you were already in, then removed. Now we keep a count
of the times you're in the room. Fixes #930