Eliminate LastAnswer synthetic collection
Generate "that was easy" on the client side by observing the puzzles table. The old way probably mattered when a) puzzles and metas were in different tables, and b) we didn't necessarily subscribe to all puzzles all the time, but neither of those are now true; this pattern will be easier to add the sound effect for partial answers to.