Run app as a dedicated user (#1498)
Running as nobody/nogroup is discouraged, so change the install script
to create a user and group, and the Systemd units to run as it. This
will also make it practical to log in for application default
credentials as a human account, which may be necessary for some users as
new service accounts don't get Drive quota.