Clean up some indices on the Message collection (#1499)
{to, room_name, timestamp} and {nick, room_name, timestamp} were added
with the expectection that the room subscription would use them, but
Mongodb prefers to use the use the {room_name, timestamp} index and
filter out the few private messages, which makes sense because they're
rare.

{timestamp} was only used by the Hubot observer, but I changed that to
use Mongo's native change stream (to not keep the messages in memory) so
it's no longer used.

{mention{ should always have had a partial filter expression since we
never want to search for messages that don't mention anybody.

Fixes #1435